:wrench: QA: fix coveralls not merging the results
By the looks of it, something changed in Coveralls in September which is causing all files to be listed 6 x times and the total coverage calculation to be off.

This should improve things a bit. Still not working 100% correctly as all files are still listed 6 times, but at least the total now seems to be calculated correctly.

Example problem report: https://coveralls.io/builds/27804200 (708 files)
Last good report I could find from mid September 2019: https://coveralls.io/builds/25691592 (118 files <- this is the correct number of files)

Refs:
* https://docs.coveralls.io/parallel-build-webhook
* https://github.com/lemurheavy/coveralls-public/issues/1390
* https://github.com/lemurheavy/coveralls-public/issues/1362
* https://github.com/lemurheavy/coveralls-public/issues/1360
* https://github.com/lemurheavy/coveralls-public/issues/1359